Task #4820: Merged trunk into branch, resolved conflict in MACProcessScope

+++ branches/Cobalt-Task4820/RTCP/Cobalt/CoInterface/test/tTABTranspose.cc
@@ -446,7 +446,7 @@ SUITE(MultiReceiver) {
 #     pragma omp section
       {
 #       pragma omp parallel for num_threads(nrReceivers)
-        for (int r = 0; r < nrReceivers; ++r) {
+        for (size_t r = 0; r < nrReceivers; ++r) {
           LOG_DEBUG_STR("Receiver thread " << r);
 
           // Set up pool where all data ends up
@@ -505,13 +505,13 @@ SUITE(MultiReceiver) {
       {
 
 #       pragma omp parallel for num_threads(nrSenders)
-        for (int s = 0; s < nrSenders; ++s) {
+        for (size_t s = 0; s < nrSenders; ++s) {
           LOG_DEBUG_STR("Sender thread " << s);
 
           MultiSender::HostMap hostMap;
 
           for (int t = 0; t < nrTABs; ++t) {
-            int r = t % nrReceivers;
+            size_t r = t % nrReceivers;
 
             struct MultiSender::Host host;
